Refactor: Move shared components under a unified directory
Consolidate all shared components into a single 'components' directory for better organization and maintainability. Updated all relevant import paths to reflect the new structure.
