+++ frontend/src/app/shared/components/layout/layout.ts
@@ -3,8 +3,8 @@
 import { Component } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { RouterOutlet } from "@angular/router";
-import { NavBar } from "../../navbar";
-import { SideNav } from "../../sidenav/sidenav";
+import { NavBar } from "../navbar";
+import { SideNav } from "../sidenav/sidenav";
 import { SidenavStateService } from "../../services/sidenav-state.service";
 
 @Component({
