+++ frontend/src/app/shared/components/navbar/_navbar-theme.scss
@@ -1,7 +1,7 @@
 // src/app/shared/navbar/_navbar-theme.scss
 
 @use "sass:map";
-@use "@angular/material" as mat;
+@use "../../../../../node_modules/@angular/material/index" as mat;
 
 @mixin theme($theme) {
     $primary: map.get($theme, primary);
