+++ frontend/src/app/shared/components/navbar/navbar.scss
