+++ frontend/src/app/shared/components/sidenav/sidenav.html
@@ -13,7 +13,7 @@
                 <img
                     class="logo sidenav-icon"
                     matButtonIcon
-                    src="../../../assets/img/site/pan-orange.svg"
+                    src="../../../../assets/img/site/pan-orange.svg"
                     alt="angular"
                     aria-hidden="true"
                 />
