+++ frontend/src/app/shared/components/sidenav/sidenav.scss
