+++ frontend/src/app/shared/components/sidenav/sidenav.ts
@@ -7,7 +7,7 @@ import { MatListModule } from "@angular/material/list";
 import { MatIconModule } from "@angular/material/icon";
 import { MatDividerModule } from "@angular/material/divider";
 import { RouterModule } from "@angular/router";
-import { SidenavStateService } from "../services/sidenav-state.service";
+import { SidenavStateService } from "../../services/sidenav-state.service";
 import {
     animate,
     state,
