+++ frontend/src/app/shared/components/style-manager/index.ts
