+++ frontend/src/app/shared/components/style-manager/style-manager.ts
