+++ frontend/src/app/shared/components/svg-viewer/svg-viewer.ts
