+++ frontend/src/app/shared/components/theme-picker/theme-picker.ts
